Create first version of Strudy out of Reffy
This moves the spec analysis logic from Reffy to a dedicated repo, as envisioned
in https://github.com/w3c/reffy/issues/747.

This also creates a CLI that can create the same reports as those that Reffy
used to create.

A few notes on things not done yet:
- This makes no attempt at improving the analysis logic that existed in Reffy.
As much as possible, the code was simply copy-and-pasted.
- No tests for now for the same reason: there weren't any tests on that part in
Reffy.
- Some of the options remain clunky. I would expect various breaking changes on
this project, time permitting.
- The code depends on code exposed by Reffy, notably the `expandCrawlResults`
and the `isLatestLevelThatPasses` functions. Reffy should export them more
directly.
- The code also currently depends on Reffy's `canonicalize-url.mjs` module and
the import is rather clumsy for now. To be fixed later on.
- The code references CSS and JS currently in Webref which should be moved here
or elsewhere.

A couple of additional analysis CLIs were extracted from Reffy:
`check-missing-dfns` and `study-backrefs`. Same thing, this is a copy-and-paste
for now.