Re-structure Web IDL report per anomaly (#325)
The spec at the root of a Web IDL cannot always be determined. The anomalies
report is now structure as a list of anomalies, and each anomaly is associated
with one or more specs.

The object structure for anomalies will likely evolve over time, e.g. to make it
easy for consumers to serialize a message with links.

This update also fixes or improves some of the analyses, notably those around
exposure.

Tests adjusted and completed accordingly.