Expand checks run by Web IDL study tool
This expands the Web IDL study tool to run additional checks:
- Report unknown types
- Report unknown extended attributes
- Includes statements: target needs to be of type "interface"
- Includes statements: mixin source (includes) needs to exist
- Includes statements: type of mixin source needs to be "interface mixin"
- Inheritance: same type required

These checks match (or should match :)) those performed in Webref already.

Reported anomaly objects could be further improved, e.g. to provide a more
complete reference to the problematic IDL when it is at the attribute or
parameter level and not at the definition level. To be done over time when
we start submitting actual anomaly reports!

This would for example be useful in the case where an interface definition
"Foo" defines 5 `bool` attributes (instead of `boolean`). The tool will
currently report 5 anomalies, all saying that the interface "Foo" references
unknown type `bool`.

Still missing:
- Split unknown types into unknown and "known unknown" types, e.g. when an
interface references a namespace
- Checks once IDL gets merged: overloaded and duplicate methods