Complete Web IDL analysis of unknown types
The IDL may reference objects that look like types but that are not types,
namely `namespace` and `interface mixin` constructs which, according to the
Web IDL spec do not create types. The tool now reports a new `wrongType` anomaly
in that case.

Note commit also sorts the list of possible anomalies to ease maintenance.