Web IDL: Check overloaded operations and redefined members
This completes the Web IDL study tool with checks to report on invalid
overloaded operations and duplicated members within interface definitions but
also across partials, mixins and partial mixins. Two new types of anomalies get
reported: `overloaded` and `redefinedMember`.

As opposed to the test logic in Webref, no actual merge is being performed.
Definitions are rather compared one by one, which makes it possible to report
more fine-grained anomaly messages and in particular target the spec that
likely needs fixing.

Typically, when a partial dfn introduces a conflict with a base dfn, it seems
logic to blame the partial dfn. Similarly, when a dfn introduces a conflict with
one of its mixins, it seems logic to blame the dfn, since the dfn includes the
mixin and not the other way round.

Only cases when multiple specs need to be blamed is when two specs create
overloads in partial dfns (no a priori reason to consider that a partial has
priority over another partial).

Note that the code will create multiple similar anomalies when the same method
is overloaded in more than 2 partial dfns and in more than 2 specs. That is,
theoretically, report could contain only one anomaly for the overloaded
operation, linked to all specs that define the overload, but in practice report
will contain one anomaly per pair of specs. That's not wrong per se, just not
optimized. Such situations should not happen much in practice, if at all, unless
the IDL starts being very clunky.

This update also rewrites tests to check on the number of anomalies first and
report a dump of the report or anomaly when assertion fails. This is to ease
debugging when a test fails.