Integrate feedback on test suite
New more specific assert methods defined in particular. One to check the number
of anomalies received, and one to check that an anomaly object "matches" the
expected anomaly object, in the sense that it contains the expected properties
with "matching" values for these properties.

The lodash library is no longer needed as a result.