Compute multi-spec repositories automatically (#660)
The list of multi-spec repositories was hardcoded and slightly outdated. The
update now computes that list automatically from the data present in the
results of the crawl (which comes from browser-specs). Compared to the previous
list, the new logic misses `w3c/woff` because it voluntarily skips repositories
that contain multiple versions of the same spec to avoid false positives. It
adds:

https://github.com/httpwg/http-extensions
https://github.com/immersive-web/real-world-geometry
https://github.com/w3c/aria
https://github.com/w3c/encrypted-media
https://github.com/w3c/gamepad
https://github.com/w3c/reporting
https://github.com/w3c/webcodecs
https://github.com/WebAssembly/threads
https://github.com/WebBluetoothCG/web-bluetooth
https://github.com/WICG/nav-speculation
https://github.com/WICG/shape-detection-api
https://github.com/WICG/WebApiDevice

Some of them contain one "main" spec and side specs that we will probably
ignore for some time, but then it does not hurt to prefix issues with the
spec's shortname in such cases, even if the targeted spec is somewhat obvious.