Rewrite file-issue-for-review logic
The script used to run the study *and* handle Git and GitHub commands. It now
expects that some analysis got run with the main strudy CLI, and only handles
Git and GitHub commands.

One update is that the script can also create pull requests for issue files
that got deleted during the analysis (because study did not reveal any issue
with the underlying spec anymore).