Delete previous CLIs and previous "study crawl" logic
Delete all the things! The previous CLIs should no longer be needed as the
main Strudy CLI now should take care of everything.

The `study-crawl` CLI (and companion `generate-report`) is a sort of exception
to the rule: all anomalies it reported are now covered by the CLI, but the
CLI could also output a dependencies report, compute a diff between two
reports, and generate an HTML report. New Strudy CLI no longer can do any of
that. This update drops the files anyway. We could perhaps keep them around,
but we haven't maintained the code, and `w3c/webref-analysis` is, I think, the
only project that uses it and I propose to shelve that project now that we
have, at least theoretically, a more flexible way to create issue reports.

The NPM package would only export the main `study` function, because there
shouldn't be any need to export the individual study functions such as
`studyWebIdl` anymore. We'll have to update Webref tests accordingly.

Project dependencies adjusted to drop `node-pandoc`... and `node-fetch` which
was no longer being used in any case.