Merge missing task anomalies and add guidance (#675)
This adds a generic mechanism to give additional guidance at the end of an
issue file, typically useful to link to the relevant section in the HTML spec
for missing "queue a global task" problems.

Guidance may also be set at the anomaly group level if we ever need that.

To avoid reporting the same guidance twice, I merged the two types of missing
task anomalies. That also allows us to reuse the same `type+spec` structure as
that used for broken links and discontinued references.

(Minor side fix: the initial run of a CLI may take some time and `this.slow` in
Mocha does not change the 2s timeout. Explicit call to `this.timeout` added.)