Add --sort option to CLI
Specs are sorted per URL in browser-specs. That order is not fantastic
when preparing a human-readable report such as the one in:
https://github.com/w3c/strudy/issues/663

The --sort option lets makes it possible to specify the order to follow
at each level of the structured report: one of "default", "name" or
"title". We may want to add additional possibilities over time.