Major release of Strudy - v3.0.0
**Breaking changes: basically everything!**

The CLI was almost entirely re-written. It can no longer produce a
generic HTML report. Instead, it can create a flexible set of reports
structured per spec, anomaly type or group, depending on parameters.
The CLI can also generate and update issue files that can then be
reviewed and submitted to the GitHub repository that hosts the
underlying spec.

The CLI command `inspect` needs to be specified to get the CLI to work,
run `npx strudy inspect --help` for details.

On top of the CLI, the new `study` function is now the only function
that gets exported for use as a library in external projects. Similar
results as before can be produced by tweaking function parameters,
notably the `what` and `structure` options.

Also, the code is now ESM-only.

Main other changes:
- Dependencies bumped to their latest version
- New analysis available for missing tasks in algorithms
- Several fixes for the broken links analysis
- Repositories that contain more than one spec are now automatically
computed, looking into the list of specs in browser-specs.