Fix "flat" report structure (#679)
With a "flat" report, the `looksGood` property can but be an empty list.
This had been left out in the code, so no way to get a "flat" report
anymore.

The reporting of a "flat" structure also needed to be improved to return
the spec(s) along with the anomaly message. The code only reported the
anomaly message, meaning that there was no way to know where the anomaly
had been detected.