[Issue filer] Use `git status` to list issue files to report
As raised in #680, `git diff` does not report issue files that still need to be
added. `git ls-files` would report these files (with the `--others` option) but
there would be no way to differentiate between modified files and deleted
files. This update rather fixes the problem by using `git status` with the
`--porcelain` option to get an easy-to-parse format.

Added bonus, this makes it possible to differentiate between "add", "update"
and "delete" actions. It's also easier to detect situations where the local
issues folder contains other types of changes that got introduced outside of
the usual workflow and that would better be handled by a human being.

Fix #680.