[Issue filer] Start new branches from initial one
The code created new branches out of the last created branch, which
effectively meant that the commits built up from one report to the next.
This update specifies that the starting point for creating a new branch
should always be the initial branch (from which we computed the list of
changes).