New `--cc` option to copy people in issues
The CLI now accepts a `--cc` option that can take a list of GitHub
handles. It will add a Cc message with these handles at the end of
issues when these issues may warrant more discussion (only the case
for algorithms and Web IDL anomalies for now).

The option is only meaningful if the --issues option is set.

To avoid hardcoding names in code, a new `CC` variable was added to
the repository, initially set to "dontcallmedom tidoust", so as to end up with:

```
<sub>Cc @dontcallmedom @tidoust</sub>
```