Web IDL study: detect event handlers with no matching even
When an interface defines an event handler, there should always be an event
named after the event handler that targets the interface. The new `noEvent`
anomaly detects situations where the event is missing.

The analysis reports the two anomalies mentioned in:
https://github.com/w3c/webref/issues/1216#issuecomment-2068997553

(The analysis reports additional anomalies when run on the raw extracts. That's
normal and due to missing event targets that get added during curation)