Add CDDL validation using cddlparser and pyodide (#837)
There is no immediate parser in Node.js that could be used to validate CDDL
extracts. Instead, the idea is to run the cddlparser Python package using a
Python interpreter running in WebAssembly.

This provides a CDDL validation guarantee for Webref.

Some notes:
- The version of cddlparser used by Strudy is specified in the
`requirements.txt` file. Goal is for Dependabot to create a pull request
automatically when a new version becomes available.
- Loading could be optimized not to depend on `micropip` at all, but that
requires hardcoding the Wheel URL of the cddlparser package somehow.
- The line position in the anomalies report needs to be adjusted to account for
the header that Reffy adds and that gets stripped during "expansion". That's
hardcoded for now. I'd say that's good enough...
- The CDDL parser stops as soon as it encounters an error, the list of
anomalies only contains the first problem as a result.
- The CDDL parser does not validate additional CDDL semantics for now.