Relax "queue a task" regexp for custom tasks
The regular expression used to detect "queue a [custom] task" only accounted
for cases where "[custom]" contained at most 2 words. This update relaxes the
regular expression to account for situations where it contains 3 words, as
happens in the Handwriting Recognition API.

The regular expression remains somewhat strict on purpose not to match on too
many things for now. We may need to relax it later on, I just propose to adjust
it as we learn from experience.