Add initial CSS anomalies tests (#980)
This creates a new module that checks CSS extracts. New anomalies reported
when:

1. A CSS construct that looks like a type (resp. function) is defined as a
function (resp. type), e.g., if `foo()` is defined as `type`. That should not
happen much. Last anomalies of this type were fixed recently in css-values-5.
2. A CSS construct gets redefined in unrelated specs. CSS properties were
already checked in Webref. This extends the check to all types of constructs.
This reports a number of duplicates when run against raw crawl results. That's
normal, duplicates are handled through curation. Two additional type dfns get
reported in css-color-5 and css-color-hdr. Some patch will be needed to address
them.
3. A CSS construct is extended in a spec, but no base definition seems to
exist. This was already checked by Webref. No problems detected for now.
4. A CSS definition was found without any link back to the actual definition
in the spec. This was also already checked by Webref. No problems detected
for now.
5. A CSS construct with a definition scoped to another construct also has an
unscoped definition. About 10 occurrences today but some of them are due to
the unscoped dfn appearing in a spec that has not yet been published to /TR,
so I would recommend ignoring them for now.

Goal is to update the tests in Webref to rather use Strudy and extend the
guarantees accordingly (excluding 5 above initially).


Co-authored-by: Dominique Hazael-Massieux <dom@w3.org>